requirements.txt
requests>=2.32.0
tqdm>=4.66.0
python-dotenv>=1.0.0
openai>=1.0.0
streamlit>=1.28.0
